Undid previous commit, added logic for file naming
The only time --use_sfftools was passed to process_sff was
for the list of special cases in
run_process_sff_through_split_lib.py, and it is needed there
to use --notrim. The upshot is gzipped SFFs now work in all
cases UNLESS that SFF is in the special cases list.