Add a mechanism to contact us
Add a function to send an email under the utils module, note that the
__init__.py file had to be added.

error.psp, now handles different type of errors that the user might face
through the workflow.

help_request.psp allows the user to send an e-mail to contact us, this
page has to be revised and formatted by @meganap

index.psp adds 'troubleshooting' as a key to hold any type of useful
information. the usage of this key can be found in
process_verification.psp and in verification.psp

This should close #168