Change study summary display
This select box will allow you now to show the study summary when
navigating using the arrow keys as well as clicking on the study.

The number of studies you can select has now been limited to one
previously it was allowing you to select multiple studies and making
the interaction confusing.

The callback javascript function was changed to only take one argument
the id of the select box so it retrieves the selected item and displays
the study summary data.