Add instructions in spanish for intl. shipping
A new link should appear on the left sidebar leading you to instructions
in spanish, the diagram synthesizing the instructions should also appear
with labels in spanish.

international_shipping.psp has been modified to take a new GET var, this
variable being the language you want the instructions to be displayed
in.