Fix #547
Made chages to form_complete.js so that the ontology short name is
returned when an ontology term is selected. Preveiously http was
returned I also added comments about what is coming back from bioportal
today so we know what are starting point was if anything changes in the
future.
I also fixed an obvious logic flaw in ontology.js