fixes #561
This should be good to go with all the data on the test server.
Before putting it on production i need to add the column to the
database and update the stored procedure.
Also addresses #559 the userid portion is removed so we can see
if we want to remove the code or leave it in.