First attempt at automatic password reset
It prints the email message if cannot send an email
so you can copy the link into your browser when using the
virtual machine.

Right now for some reason I have to wait about 20-30 seconds
before the link with the code to change the password is valid
I'm not sure what that is about, I could add a sleep statement if
we think it will be a problem.